# openapi.yaml
openapi: 3.0.3
info:
  title: OAuth 2.0 Dummy Endpoint
  version: 1.1.0
  description: |
    Minimal OAuth 2.0 authorization server for demos and integration testing.
    Implements the Authorization Code grant (RFC 6749 section 4.1) with PKCE
    (RFC 7636), the Client Credentials grant (RFC 6749 section 4.4), and the
    Refresh Token grant (RFC 6749 section 6). Issues signed JWT access tokens.
    Also supports introspection (RFC 7662) and revocation (RFC 7009).
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://oauth2-dummy-endpoint.onrender.com
    description: render.com deployment (replace with your service URL)
  - url: http://localhost:3000
    description: Local development
tags:
  - name: OAuth
    description: OAuth 2.0 endpoints
  - name: Discovery
    description: Authorization server metadata
  - name: Ops
    description: Health and documentation
paths:
  /oauth/authorize:
    get:
      tags:
        - OAuth
      summary: Authorization endpoint (RFC 6749 section 4.1.1)
      operationId: authorize
      description: |
        Starts the Authorization Code flow. The browser-facing endpoint that
        renders a consent screen and, on approval, redirects back to the
        client with an authorization code.
      parameters:
        - name: response_type
          in: query
          required: true
          schema:
            type: string
            enum:
              - code
        - name: client_id
          in: query
          required: true
          schema:
            type: string
          example: demo-client
        - name: redirect_uri
          in: query
          required: true
          schema:
            type: string
            format: uri
          example: http://localhost:3000/callback
        - name: scope
          in: query
          required: false
          schema:
            type: string
          example: read write
        - name: state
          in: query
          required: false
          description: Opaque value for CSRF protection; echoed back on redirect.
          schema:
            type: string
        - name: code_challenge
          in: query
          required: false
          description: PKCE code challenge (RFC 7636).
          schema:
            type: string
        - name: code_challenge_method
          in: query
          required: false
          schema:
            type: string
            enum:
              - plain
              - S256
      responses:
        '200':
          description: Consent HTML page rendered to the end user.
          content:
            text/html:
              schema:
                type: string
        '302':
          description: |
            Redirect back to redirect_uri with either ?code=...&state=... on
            approval, or ?error=access_denied&state=... on denial.
          headers:
            Location:
              schema:
                type: string
                format: uri
        '400':
          description: Invalid request (unknown client or redirect_uri mismatch).
          content:
            text/html:
              schema:
                type: string
  /oauth/token:
    post:
      tags:
        - OAuth
      summary: Token endpoint
      operationId: issueToken
      description: |
        Exchanges credentials for an access token. Supports grant types
        `authorization_code`, `client_credentials`, and `refresh_token`.
        Client credentials may be sent as HTTP Basic or as form fields.
      security:
        - clientBasic: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/TokenRequest'
      responses:
        '200':
          description: Access token issued.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Invalid request or grant.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthError'
        '401':
          description: Invalid client credentials.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthError'
  /oauth/introspect:
    post:
      tags:
        - OAuth
      summary: Introspect an access token (RFC 7662)
      operationId: introspectToken
      security:
        - clientBasic: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/IntrospectionRequest'
      responses:
        '200':
          description: Introspection result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntrospectionResponse'
        '401':
          description: Invalid client credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthError'
  /oauth/revoke:
    post:
      tags:
        - OAuth
      summary: Revoke an access or refresh token (RFC 7009)
      operationId: revokeToken
      security:
        - clientBasic: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/RevocationRequest'
      responses:
        '200':
          description: Token revoked. Returns 200 even for unknown tokens per RFC 7009.
        '401':
          description: Invalid client credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthError'
  /oauth/userinfo:
    get:
      tags:
        - OAuth
      summary: Return claims for the bearer token
      operationId: getUserInfo
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Token claims
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
        '401':
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthError'
  /.well-known/oauth-authorization-server:
    get:
      tags:
        - Discovery
      summary: Authorization server metadata (RFC 8414)
      operationId: getServerMetadata
      responses:
        '200':
          description: Metadata document
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorizationServerMetadata'
  /healthz:
    get:
      tags:
        - Ops
      summary: Liveness probe
      operationId: getHealth
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
components:
  securitySchemes:
    clientBasic:
      type: http
      scheme: basic
      description: HTTP Basic auth with client_id and client_secret.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token issued by /oauth/token.
    oauth2AuthorizationCode:
      type: oauth2
      description: OAuth 2.0 Authorization Code flow (PKCE supported).
      flows:
        authorizationCode:
          authorizationUrl: https://oauth2-dummy-endpoint.onrender.com/oauth/authorize
          tokenUrl: https://oauth2-dummy-endpoint.onrender.com/oauth/token
          refreshUrl: https://oauth2-dummy-endpoint.onrender.com/oauth/token
          scopes:
            read: Read access
            write: Write access
            openid: OpenID Connect subject identifier
            profile: Basic profile info
            email: Email address
    oauth2ClientCredentials:
      type: oauth2
      description: OAuth 2.0 Client Credentials flow.
      flows:
        clientCredentials:
          tokenUrl: https://oauth2-dummy-endpoint.onrender.com/oauth/token
          scopes:
            read: Read access
            write: Write access
  schemas:
    TokenRequest:
      type: object
      required:
        - grant_type
      properties:
        grant_type:
          type: string
          enum:
            - authorization_code
            - client_credentials
            - refresh_token
          example: authorization_code
        code:
          type: string
          description: Required when grant_type=authorization_code. The code from /oauth/authorize.
        redirect_uri:
          type: string
          format: uri
          description: Required when grant_type=authorization_code. Must match the value used at /oauth/authorize.
        code_verifier:
          type: string
          description: PKCE verifier. Required if a code_challenge was sent to /oauth/authorize.
        refresh_token:
          type: string
          description: Required when grant_type=refresh_token.
        scope:
          type: string
          description: Space-delimited list of scopes.
          example: read write
        client_id:
          type: string
          description: Required if not using HTTP Basic auth.
          example: demo-client
        client_secret:
          type: string
          description: Required if not using HTTP Basic auth (confidential clients).
          example: demo-secret
    TokenResponse:
      type: object
      required:
        - access_token
        - token_type
        - expires_in
      properties:
        access_token:
          type: string
          description: Signed JWT (HS256).
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6ImF0K2p3dCJ9.eyJzdWIiOiJkZW1vLXVzZXItMDAxIn0.signature
        token_type:
          type: string
          enum:
            - Bearer
          example: Bearer
        expires_in:
          type: integer
          format: int32
          description: Seconds until access token expiry.
          example: 3600
        refresh_token:
          type: string
          description: Returned only for authorization_code grant.
          example: aBcDeF1234567890_refreshTokenValue
        scope:
          type: string
          example: read write
    IntrospectionRequest:
      type: object
      required:
        - token
      properties:
        token:
          type: string
        token_type_hint:
          type: string
          enum:
            - access_token
            - refresh_token
        client_id:
          type: string
        client_secret:
          type: string
    IntrospectionResponse:
      type: object
      required:
        - active
      properties:
        active:
          type: boolean
          example: true
        scope:
          type: string
          example: read write
        client_id:
          type: string
          example: demo-client
        token_type:
          type: string
          example: Bearer
        exp:
          type: integer
          format: int64
          example: 1717440000
        iat:
          type: integer
          format: int64
          example: 1717436400
        sub:
          type: string
          example: demo-user-001
        aud:
          type: string
          example: https://oauth2-dummy-endpoint.onrender.com
        iss:
          type: string
          example: https://oauth2-dummy-endpoint.onrender.com
        jti:
          type: string
          example: 6f2a3b4c-5d6e-7f80-9a0b-1c2d3e4f5a6b
    RevocationRequest:
      type: object
      required:
        - token
      properties:
        token:
          type: string
        token_type_hint:
          type: string
          enum:
            - access_token
            - refresh_token
    UserInfo:
      type: object
      properties:
        sub:
          type: string
          example: demo-user-001
        client_id:
          type: string
          example: demo-client
        scope:
          type: string
          example: read write
        email:
          type: string
          format: email
          example: demo-user@example.com
        name:
          type: string
          example: Demo User
        iss:
          type: string
          example: https://oauth2-dummy-endpoint.onrender.com
        aud:
          type: string
          example: https://oauth2-dummy-endpoint.onrender.com
        exp:
          type: integer
          format: int64
          example: 1717440000
    AuthorizationServerMetadata:
      type: object
      properties:
        issuer:
          type: string
          format: uri
        authorization_endpoint:
          type: string
          format: uri
        token_endpoint:
          type: string
          format: uri
        introspection_endpoint:
          type: string
          format: uri
        revocation_endpoint:
          type: string
          format: uri
        userinfo_endpoint:
          type: string
          format: uri
        grant_types_supported:
          type: array
          items:
            type: string
        response_types_supported:
          type: array
          items:
            type: string
        token_endpoint_auth_methods_supported:
          type: array
          items:
            type: string
        code_challenge_methods_supported:
          type: array
          items:
            type: string
        scopes_supported:
          type: array
          items:
            type: string
        id_token_signing_alg_values_supported:
          type: array
          items:
            type: string
        service_documentation:
          type: string
          format: uri
    HealthResponse:
      type: object
      properties:
        status:
          type: string
          example: ok
    OAuthError:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          enum:
            - invalid_request
            - invalid_client
            - invalid_grant
            - unauthorized_client
            - unsupported_grant_type
            - invalid_scope
            - invalid_token
            - access_denied
            - not_found
          example: invalid_grant
        error_description:
          type: string
          example: Unknown or used code